Backed out changeset 16c9098dd3d2.
See http://mid.gmane.org/8763jbsyr7.fsf@xemacs.org for details from Stephen
as to why. Basically, there was a bug in his version of Xt that broke the
initialisation of the describe-char-unicodedata-file variable; he has found
it, and fixed it locally, so this should not longer be an issue.